25/11/2022

+++ src/components/base/dialog/dialogDelete.vue
@@ -3,16 +3,19 @@
         <div class="m-noti-box">
             <div class="box-content">
                 <div class="content">
-                    <div class="icon-delete"></div>
-                    <div class="message-content">
-                        Bạn có thực sự muốn xóa nhân viên "{{ employeeName }}"
-                        không?
+                    <div class="title-noti">
+                        Xóa{{ isDeleteBatch ? ` ${listClicked.length}` : "" }}
+                        người dùng
                     </div>
+                    <div class="icon-close" @click="oClDialogDelete"></div>
+                </div>
+                <div class="message-content">
+                    Bạn có chắc muốn xóa người dùng đã chọn không?
                 </div>
                 <div class="content-footer">
-                    <div>
+                    <div class="mr-8">
                         <ms-button
-                            btnText="Hủy"
+                            btnText="Không"
                             btnExtra
                             @click="oClDialogDelete"
                         ></ms-button>
@@ -20,7 +23,7 @@
                     <div class="content-footer-left">
                         <ms-button
                             class="btnToastDelete"
-                            btnText="Đồng ý"
+                            btnText="Có, xóa người dùng"
                             @click="handleDelete(employeeID)"
                         ></ms-button>
                     </div>
@@ -29,24 +32,28 @@
         </div>
         <ms-loading v-show="isLoading"></ms-loading>
         <div class="overlay" @click="oClDialogDelete"></div>
-        <!-- <ms-toast v-show="openToast" toastAct="xóa"></ms-toast> -->
     </div>
 </template>
 <script>
 import axios from "axios";
 import MsButton from "../button/MsButton.vue";
 import { BASE_URL } from "../../../constans/constans";
-import MsLoading from "../loading/MsLoading.vue";
 
 export default {
-    components: { MsButton, MsLoading },
+    components: { MsButton },
     data() {
         return {
-            openToast: true,
             isLoading: false,
         };
     },
-    props: ["oClDialogDelete", "employeeID", "employeeName", "reloadList"],
+    props: [
+        "oClDialogDelete",
+        "employeeID",
+        "employeeName",
+        "reloadList",
+        "listClicked",
+        "isDeleteBatch",
+    ],
     methods: {
         /**
          * Thực hiện xóa nhân viên theo ID
@@ -55,13 +62,28 @@ export default {
         async handleDelete(employee) {
             try {
                 this.isLoading = true;
-                const res = await axios.delete(`${BASE_URL}/${employee}`);
-                if (res) {
-                    this.openToast = true;
-                    this.isLoading = false;
-                    this.reloadList();
-                    this.oClDialogDelete();
-                    this.$emit("openToast", this.openToast);
+                if (this.listClicked.length > 0) {
+                    axios
+                        .post(`${BASE_URL}/deleteBatch`, {
+                            EmployeeIDs: this.listClicked,
+                        })
+                        .then(res => {
+                            this.isLoading = false;
+                            this.reloadList();
+                            this.oClDialogDelete();
+                            this.$emit("openToast", true);
+                        })
+                        .catch(e => {
+                            console.log(e);
+                        });
+                } else {
+                    const res = await axios.delete(`${BASE_URL}/${employee}`);
+                    if (res) {
+                        this.isLoading = false;
+                        this.reloadList();
+                        this.oClDialogDelete();
+                        this.$emit("openToast", true);
+                    }
                 }
             } catch (error) {
                 console.log(error);
@@ -73,7 +95,11 @@ export default {
          * Author: NHAnh (06/11/2022)
          */
         stopLoading(e) {
-            this.isLoading = !this.isLoading;
+            try {
+                this.isLoading = !this.isLoading;
+            } catch (err) {
+                console.log(err);
+            }
         },
     },
 };
@@ -92,43 +118,43 @@ export default {
     z-index: 999;
 }
 .m-noti-box {
-    width: 444px;
-    min-width: 444px;
-    height: 183px;
+    width: 400px;
+    min-width: 400px;
     background-color: #fff;
     border-radius: 4px;
     z-index: 1;
 }
 .box-content {
-    padding: 32px;
+    padding: 24px;
     height: 100%;
     box-sizing: border-box;
 }
 .content {
     display: flex;
-    align-items: flex-start;
+    justify-content: space-between;
+}
 
-    border-bottom: 1px solid #b8bcc3;
+.title-noti {
+    font-weight: 700;
+    font-size: 20px;
 }
 .content-footer {
     margin-top: 20px;
     display: flex;
-    justify-content: space-between;
+    justify-content: flex-end;
 }
 .content-footer-left {
     display: flex;
 }
 .message-content {
     overflow: auto;
-    max-height: 400px;
-    margin-bottom: 32px;
-    padding-left: 16px;
-    height: 37px;
+    margin-top: 24px;
+    margin-bottom: 40px;
     display: flex;
     align-items: center;
 }
 .btnToastDelete {
-    background-color: #e60000;
+    background-color: #e81e1e;
 }
 .btnToastDelete:hover {
     background-color: #eb3333;
